[chore][az] removes comments and renames files
