+++ src/services/AuthService.ts
@@ -1,6 +1,5 @@
-// src/services/authService.ts
 import axios from 'axios'
-import { FormState } from '@/types/LoginInterface'
+import { FormState } from '@/types/AuthInterface'
 
 const API_BASE_URL = import.meta.env.VITE_API_BASE_URL
